% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_estimators.R
\name{g}
\alias{g}
\title{Creates an individual term in an estimator}
\usage{
g(fs)
}
\arguments{
\item{fs}{a \code{list} of \code{function}s of the form \code{function(vs) { function(...) {}}},
where \code{vs} is a \code{list} of data elements the may be used in the inner function.}
}
\value{
a partial function of a single argument \code{vs}
}
\description{
Creates an individual term in an estimator
}
