% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_estimators.R
\name{h}
\alias{h}
\title{Creates a linear estimator}
\usage{
h(aFUN, bFUN, post = ratio, collector)
}
\arguments{
\item{aFUN}{a function returned by \href{`g`}{\code{g}}}

\item{bFUN}{a function returned by \href{`g`}{\code{g}}}

\item{post}{a function applied to the final result of the pair returned by
the  \code{collector} function. Defaults to \href{`ratio`}{\code{ratio}}.}

\item{collector}{the function used to collapse the monoidal data.}
}
\value{
a partial function of a single argument \code{vsl}: a \code{list} of \code{list}s
}
\description{
Creates a linear estimator
}
