% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_estimators.R
\name{make_estimator}
\alias{make_estimator}
\title{Creates an estimator}
\usage{
make_estimator(as, b)
}
\arguments{
\item{as}{a \code{list} of a functions}

\item{b}{a single function}
}
\description{
Creates an estimator
}
