% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoid.R
\name{monoidal_applicators}
\alias{monoidal_applicators}
\alias{apply_prod}
\alias{apply_sum}
\alias{apply2_sum}
\title{Monoidal applicators}
\usage{
apply_prod(l, r, ...)

apply_sum(l, r, ...)

apply2_sum(l, r, ...)
}
\arguments{
\item{l}{a value of the same type returned by:}

\item{r}{a function that returns the same type as \code{l}}

\item{...}{additional arguments passed to \code{r(...)}}
}
\description{
Monoidal applicators
}
