% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoid.R
\name{monoidal_collectors}
\alias{monoidal_collectors}
\alias{collect_prod}
\alias{collect_sum}
\title{Monoidal collectors}
\usage{
collect_prod(fs, ...)

collect_sum(fs, ...)
}
\arguments{
\item{fs}{a \code{list} of \code{function}s}

\item{...}{additional arguments passed to \code{r(...)} of the \code{applicator}}
}
\description{
Monoidal collectors
}
