% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoid.R
\name{monoidal_helpers}
\alias{monoidal_helpers}
\alias{ratio}
\alias{one}
\alias{zero}
\title{Monoidal helper functions}
\usage{
ratio(x)

one(...)

zero(...)
}
\arguments{
\item{x}{a \code{list}}

\item{...}{additional arguments passed to \code{r(...)}}
}
\description{
Monoidal helper functions
}
