% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_estimators.R
\name{h}
\alias{h}
\title{Creates a linear estimator}
\usage{
h(aFUN, bFUN, post = identity, collector = collect2_sum)
}
\arguments{
\item{aFUN}{a function returned by \href{`g`}{\code{g}}}

\item{bFUN}{a function returned by \href{`g`}{\code{g}}}

\item{post}{a function applied to the final result. Defaults to \code{identity}}

\item{collector}{the function used to collapse the monoid. Defaults to
\code{collect2_sum}.}
}
\value{
a partial function of a single argument \code{vsl}: a \code{list} of \code{list}s
}
\description{
Creates a linear estimator
}
