% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_estimators.R
\name{h}
\alias{h}
\title{Creates a linear estimator}
\usage{
h(gFUN)
}
\arguments{
\item{gFUN}{a function returned by \href{`g`}{\code{g}}}
}
\value{
a partial function of a single argument \code{vsl}: a \code{list} of \code{list}s
}
\description{
Creates a linear estimator
}
