% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_estimators.R
\name{make_estimator}
\alias{make_estimator}
\title{Creates an estimator}
\usage{
make_estimator(fs)
}
\arguments{
\item{fs}{a \code{list} of \code{function}s of the form \code{function(vs) { function(...) {}}},
where \code{vs} is a \code{list} of data elements the may be used in the inner function.}
}
\description{
Creates an estimator
}
