% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_estimators.R
\name{make_estimator}
\alias{make_estimator}
\title{Creates an estimator}
\usage{
make_estimator(as, bs, post = ratio, collector = collect2_sum)
}
\arguments{
\item{as}{a \code{list} of a functions}

\item{bs}{a \code{list} of a functions}

\item{post}{a function applied to the final result of the pair returned by
the  \code{collector} function. Defaults to \href{`ratio`}{\code{ratio}}.}

\item{collector}{the function used to collapse the monoidal data.}
}
\description{
Creates an estimator
}
