% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoid.R
\name{monoidal_collectors}
\alias{monoidal_collectors}
\alias{collect_prod}
\alias{collect_sum}
\alias{accum_sum}
\alias{collect2_sum}
\alias{accumsum_collectsum_seq}
\alias{one}
\alias{zero}
\alias{make_pull_value}
\alias{ratio}
\title{Monoidal collectors}
\usage{
collect_prod(fs, ...)

collect_sum(fs, ...)

accum_sum(fs, ...)

collect2_sum(fs, ...)

accumsum_collectsum_seq(fs, ...)

one(...)

zero(...)

make_pull_value(name)

ratio(x)
}
\arguments{
\item{fs}{a \code{list} of \code{function}s}

\item{...}{additional arguments passed to \code{r(...)} of the \code{applicator}}
}
\description{
Monoidal collectors
}
