% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoid.R
\name{monoidal_utilities}
\alias{monoidal_utilities}
\alias{make_monoidal_applicator}
\alias{make_monoidal_collector}
\title{Utilites for working with monoidal structures}
\usage{
make_monoidal_applicator(op)

make_monoidal_collector(applicator, .init)
}
\arguments{
\item{op}{a binary operator}

\item{applicator}{one of the \url{monoidal_applicators}}

\item{.init}{the unit of the monoid (e.g. 0 for \code{+} and numbers)}
}
\value{
\code{make_monoidal_applicator} returns a function of at least 2 arguments:
\itemize{
\item \code{l}: a value of the same type returned by:
\item \code{r}: a function that returns the same type as \code{l}
\item \code{...}: optional arguments passed to \code{r}
}

\code{make_monoidal_collector} returns a function of at least 1 argument:
\itemize{
\item \code{fs}: a \code{list} of functions
\item \code{...}: optional arguments passed to \code{r} in the \code{applicator}
}
}
\description{
Utilites for working with monoidal structures
}
